\name{QuantileDSAR-Data}
\alias{W}
\alias{AQIweather_df}
\alias{Economy_df}
\title{Data set}
\description{
  W is the spatial weights matrix; AQIweather_df is the data frame which includes air quality data and meteorological data; Economy_df is the data frame which includes economic data.
}
\author{Wenyu Li}

