\name{QuantileDSAR-package}
\alias{QuantileDSAR}
\docType{package}
\title{
\packageTitle{QuantileDSAR}
}
\description{
  \packageDescription{QuantileDSAR}
}
\details{
  Installation:
  
  \code{#install.packages("Rcpp")}
  
  \code{#install.packages("RcppArmadillo")}
  
  \code{#install.packages("quantreg")}
  
  \code{#install.packages("nloptr")}
  
  \code{library(Rcpp)}
  
  \code{library(RcppArmadillo)}
  
  \code{library(quantreg)}
  
  \code{library(nloptr)}
  
  \code{system("R CMD INSTALL --build QuantileDSAR_1.0.tar.gz")}
  
  \code{library(QuantileDSAR)}
}
\author{
\packageAuthor{QuantileDSAR}

Maintainer: \packageMaintainer{QuantileDSAR}
}
\references{
Spatial-temporal Model with Heterogeneous Random Effects 
}
% \keyword{ package }
% \seealso{
% ~~ Optional links to other man pages, e.g. ~~
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% ~~ simple examples of the most important functions ~~
% }
